requirements.txt
# Core Django i REST Framework (últimes versions estables)
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.5.0

# Base de dades PostGIS
psycopg[binary]==3.2.3
django-environ==0.11.2

# Servidor WSGI per producció
gunicorn==23.0.0
gevent==24.10.3

# Autenticació i seguretat
djangorestframework-simplejwt==5.3.1
django-oauth-toolkit==3.0.1
cryptography==43.0.1
argon2-cffi==23.1.0  # Per hash de contrasenyes més segur que PBKDF2

# Validació i JSON Schema
jsonschema==4.23.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Rate limiting i protecció
django-ratelimit==4.1.0
django-csp==3.8
django-defender==0.9.7  # Protecció contra brute force
django-celery-beat==2.7.0  # Per tasques programades

# Logging i monitoring
django-log-request-id==2.1.0
python-json-logger==2.0.7

# Testing i seguretat
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
bandit==1.7.10  # Anàlisi seguretat codi
safety==3.2.8  # Comprovació vulnerabilitats dependències

# Documentació API
drf-spectacular==0.27.2

# Utilitats
python-dateutil==2.9.0
pytz==2024.2
requests==2.32.3

# Processament asíncron
celery==5.4.0
redis==5.1.1

# Desenvolupament i qualitat codi
black==24.10.0
flake8==7.1.1
isort==5.13.2
pre-commit==4.0.1
mypy==1.11.2
django-stubs==5.1.0